added search to its own component
