+++ react-app/src/components/SupportPage/index.ts
@@ -1,5 +1,6 @@
 import ContactHelpDesk from "./contactHelpDesk";
 import ExpandCollapseBtn from "./expandCollapseBtn";
 import LeftNavigation from "./navigationBar";
+import SearchContent from "./SearchContent";
 
-export { ContactHelpDesk, ExpandCollapseBtn, LeftNavigation };
+export { ContactHelpDesk, ExpandCollapseBtn, LeftNavigation, SearchContent };
